+++ packages/back-end/src/integrations/Redshift.ts
@@ -19,11 +19,8 @@ export default class Redshift extends SqlIntegration {
   runQuery(sql: string) {
     return runPostgresQuery(this.params, sql);
   }
-  getFullTableName(table: string): string {
-    if (this.params.defaultSchema && !table.match(/\./)) {
-      return this.params.defaultSchema + "." + table;
-    }
-    return table;
+  getSchema(): string {
+    return this.params.defaultSchema || "";
   }
   percentile(col: string, percentile: number) {
     return `APPROXIMATE  PERCENTILE_DISC ( ${percentile} ) WITHIN GROUP (ORDER BY ${col})`;
