+++ packages/docs/pages/api-docs.mdx
@@ -4,7 +4,7 @@ The Growth Book API is pretty easy to integrate into any tech stack.
 
 To get started, on the Growth Book settings page, generate an API key.
 
-As of right now, there is only a single public API endpoint: `/config/{apiKey}`.  If you are using Managed Growth Book (https://app.growthbook.io), the endpoint is accessible from our global CDN: https://cdn.growthbook.io/config/{apiKey}.  
+As of right now, there is only a single public API endpoint: `/config/{apiKey}`.  If you are using Growth Book Cloud (https://app.growthbook.io), the endpoint is accessible from our global CDN: https://cdn.growthbook.io/config/{apiKey}.  
 
 This returns configuration info for every experiment. 
 It can be passed directly into our client libraries or you can use the info to implement your own variation assignment and targeting.
