+++ packages/front-end/components/Experiment/EditInfoForm.tsx
@@ -68,8 +68,10 @@ const EditInfoForm: FC<{
   );
   const { apiCall } = useAuth();
 
+  const datasource = getDatasourceById(value.datasource);
   const variationKeys =
-    getDatasourceById(value.datasource)?.settings?.variationIdFormat === "key";
+    (datasource?.settings?.variationIdFormat ||
+      datasource?.settings?.experiments?.variationFormat) === "key";
 
   const deleteVariation = (i: number) => {
     const variations = [...value.variations];
