+++ packages/front-end/components/Experiment/NewExperimentForm.tsx
@@ -133,8 +133,10 @@ const NewExperimentForm: FC<NewExperimentFormProps> = ({
     phases: initialPhases,
   });
 
+  const datasource = getDatasourceById(value.datasource);
   const variationKeys =
-    getDatasourceById(value.datasource)?.settings?.variationIdFormat === "key";
+    (datasource?.settings?.variationIdFormat ||
+      datasource?.settings?.experiments?.variationFormat) === "key";
 
   const deleteVariation = (i: number) => {
     const variations = [...value.variations];
